+++ bot/mrb/fun/dice.py
